Moved Lambda code examples into Lambda directory
