+++ aws/sdk/examples/lambda/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "lambda-code-examples"
+version = "0.1.0"
+authors = ["Richard H. Boyd <rhboyd@amazon.com>", "Doug Schwartz <dougsch@amazon.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+lambda = { package = "aws-sdk-lambda", path = "../../build/aws-sdk/lambda" }
+tokio = { version = "1", features = ["full"] }
+structopt = { version = "0.3", default-features = false }
+tracing-subscriber = { version = "0.2.16", features = ["fmt"] }
+aws-types = { path = "../../build/aws-sdk/aws-types" }
